Retire workspace model integration tests
