+++ Tests/QuillCodeAppTests/WorkspaceActivityIntegrationTests.swift
@@ -1,0 +1,61 @@
+import XCTest
+import QuillCodeAgent
+import QuillCodeCore
+import QuillCodeTools
+@testable import QuillCodeApp
+
+@MainActor
+final class WorkspaceActivityIntegrationTests: XCTestCase {
+    func testPlanUpdateToolRecordsNormalizedActivityPlan() throws {
+        let model = QuillCodeWorkspaceModel(activity: ActivityState(isVisible: true))
+        _ = model.newChat()
+        let update = AgentPlanUpdate(
+            explanation: "  Keep the plan visible while work proceeds.  ",
+            plan: [
+                AgentPlanItem(step: "  Inspect state  ", status: .completed),
+                AgentPlanItem(step: "Implement change", status: .inProgress, detail: "  One reviewable slice.  "),
+                AgentPlanItem(step: "Validate and summarize", status: .pending)
+            ]
+        )
+        let call = ToolCall(
+            name: ToolDefinition.planUpdate.name,
+            argumentsJSON: try JSONHelpers.encodePretty(update)
+        )
+
+        let result = model.runToolCall(call, workspaceRoot: try makeTempDirectory())
+        let decoded = try JSONHelpers.decode(AgentPlanUpdate.self, from: result.stdout)
+
+        XCTAssertTrue(result.ok, result.error ?? "")
+        XCTAssertEqual(decoded.explanation, "Keep the plan visible while work proceeds.")
+        XCTAssertEqual(decoded.plan.map(\.step), ["Inspect state", "Implement change", "Validate and summarize"])
+        XCTAssertEqual(model.selectedThread?.events.last?.summary, "\(ToolDefinition.planUpdate.name) completed")
+        XCTAssertEqual(model.surface().activity.planItems.map(\.title), [
+            "Inspect state",
+            "Implement change",
+            "Validate and summarize"
+        ])
+        XCTAssertEqual(model.surface().activity.planItems.map(\.statusLabel), ["Done", "Running", "Pending"])
+        XCTAssertEqual(model.surface().activity.planItems[1].detail, "One reviewable slice.")
+    }
+
+    func testPlanUpdateToolRejectsMultipleRunningSteps() throws {
+        let model = QuillCodeWorkspaceModel()
+        _ = model.newChat()
+        let update = AgentPlanUpdate(
+            plan: [
+                AgentPlanItem(step: "First", status: .inProgress),
+                AgentPlanItem(step: "Second", status: .inProgress)
+            ]
+        )
+        let call = ToolCall(
+            name: ToolDefinition.planUpdate.name,
+            argumentsJSON: try JSONHelpers.encodePretty(update)
+        )
+
+        let result = model.runToolCall(call, workspaceRoot: try makeTempDirectory())
+
+        XCTAssertFalse(result.ok)
+        XCTAssertEqual(result.error, "Plan update can have at most one in_progress step.")
+        XCTAssertEqual(model.root.topBar.agentStatus, "Failed")
+    }
+}
